requirements.txt
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
python-dateutil==2.8.2
tkcalendar==1.6.1
streamlit==1.35.0
plotly==5.18.0
requests==2.31.0